+++ NEWS.md
@@ -75,3 +75,5 @@
 * Fixed bug in function addCorFlex
 
 # simstudy 0.1.11
+
+* Added negative binomial option to genCorGen, #addCorGen, genCorFlex, and addCorFlex#